Changed config directory path in settings.php
The config directory path configured was '../../../../config' which was relative path from 'settings.php' file's location to configuration directory.

This created issues when drush tried to export configurations via `drush cex -y` and throws an error from FileStorage.php that it was unable to crete a config directory.

Adding `-vvv` flag to the command showed that the drush is consider the project root to be 'covid_vaccination_center/web' which means it trying to create 'config' directory outside the project as per the relative path given.

The relative path in this case should be '../config' from Drupal root which then allows the command `drush cex -y` to be successfull.