Update the limit of field_registered_users in covid center node bundle to 500.
Earlier, it was set to unlimited and idea was to add settings form for administrators/site managers to change the limit and use field hooks to get this setting and apply the limit to settings of field on the go.

However, on re-reading the requirement it was found that only a limit was required. There was no mention of site manager/administrator to control this value, so this value was added to field_registered_users limit in field settings.