Added search box on the pages for searching covid vaccination centers.
1. Added a exposed filter in all-covid-center views.
2. Exposed it as a block using better exposed filter settings.
3. Placed the block in sidebar between the register block and other cities block.
4. Changed the field type of City >> Taxonomy term >> Name from string to full text in order to make it compatible with search autocomplete
5. Configured search autocomplete on 'name' on taxonomy term in covid_center index.
6. Added a hook_form_FORM_ID_alter to remove the 'Apply' button from the search box.

The search functionaliy will work like the following.

An user will input the city name in the search box and the covid center list will start appearing as results.

As soon as the user selects the suggested results, it will be redirected to that particular covid center page.

If no suggestions are there or if user presses 'enter' button,it will be redirected to /all-covid-center page to filter the results.