Added a new field formatter for field_available_slots.
We wanted to display the available slots of a covid vaccination center on it's view page.

The available slots are reduced when an user is registered and gets increased when site administrator removes an user. Depending upon this, we needed to display the availability count.

1. We added a field formatter named AvailableSlotsFormatter which extends StringFormatter from Drupal core.
2. Added the logic of updating the available slots in viewElements().
3. Updated core.entity_form_display.node.covid_center.default config to include available slots.
4. Updated core.entity_view_display.node.covid_center.default config to include available slot
s and field_registered_users which was needed to alter the value at the time of rendering.
5. Added a new field_available_slots to covid center as string field and added it's configs.
6. Updated field_registered_user to have cardinality -1 as it will be now controlled by availability slots.