Updated the frontpage and other cities view.
After re-looking at the requirements, we realized that the other cities views was only available on city pages and front page. Along with that, we also saw that the frontpage view, which we are using in this case should only be accessible to authenticated users.

It was missing the title as per mock up as well.

Also, the frontpage view should show the vaccination center for user's own city. The other cities was also to be shown in sidebar on front page. Now, as the front page is a view that will not be providing the contextual filter, so we add it via hook_views_pre_build().

We added logic in slot_booking_customizations_views_pre_build() to get the current user city and pass it as argument in contextual filter to frontpage and other cities view.

1. Added a service helper class SlotBookingCustomizationHelper.
2. Added a function in service helper class to get the current user city.
3. Update the other cities block, other cities view and frontpage view configurations.