BONUS: Added logic to update Covid Center references if an admin changes the user covid centerand removed field_covid_center from user registeration form.
Added a hook_entity_presave() to act upon old and new covid centers selected in user entity's field_covid_center by admin or by user itself.

We don't want to have wrong data saved, that's why we are updating the new covid center selected in user entity to add the user ID in field_referenced_users and remove the same from old covid center.

We chose this hook as we can the unchanged entity by $entity->original which is used to compare new and old values. We only update the nodes (covid centers) if the covid center is actually changed in the user profile.

--------------------------------------------------------------------------------------------

We did not wanted the user to choose the covid center while registering an account on the platform, as it can only register in it's own city.

So, we extended the RegisterForm from user module and removed it if the user is anonymous.