BONUS: Allow user to register on covid center only if the slots are available.
We added a function checkSlotAvailability in RegisterButton block class to order to compare if the value in field_available_slots is more than number of users added in field_registered_users. If yes, an user is allowed to register in the covid center, otherwise not.

This check in chaining with earlier checks will make sure than an user

1. Will only be able to register in COVID center in it's city.
2. Will only be able to register if the slots are available.